мелкие обновления

+++ src/main/java/andrey/dev/backendforcursach/service/MusicService.java
@@ -2,14 +2,14 @@ package andrey.dev.backendforcursach.service;
 
 import andrey.dev.backendforcursach.dto.MusicRequest;
 import andrey.dev.backendforcursach.dto.mapper.MusicMapper;
+import andrey.dev.backendforcursach.dto.mapper.RequestMapper;
 import andrey.dev.backendforcursach.models.Music;
 import andrey.dev.backendforcursach.repositores.MusicRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 @RequiredArgsConstructor
@@ -17,6 +17,7 @@ public class MusicService {
     private final MusicRepository musicRepository;
     private final MusicMapper musicMapper;
     private final FileStorageService fileStorageService;
+    private final RequestMapper requestMapper;
 
     public void createMusic(MusicRequest musicRequest) {
         musicRepository.save(musicMapper.toMusic(musicRequest, fileStorageService));
@@ -40,10 +41,50 @@ public class MusicService {
     @Transactional
     public void updateMusic(MusicRequest musicRequest, Long id) {
         Music oldMusic = musicRepository.findById(id).orElseThrow(() -> new RuntimeException("no such music with this id"));
-        fileStorageService.deleteImageFile(oldMusic.getImgUrl());
-        fileStorageService.deleteMusicFile(oldMusic.getSongUrl());
-        Music changedMusic = musicMapper.toMusic(musicRequest, fileStorageService);
-        musicRepository.updateMusic(changedMusic, id);
+        Music musicWithoutFiles = requestMapper.toMusic(musicRequest);
+        if (musicRequest.getImg() != null) {
+            fileStorageService.deleteImageFile(oldMusic.getImgUrl());
+            musicWithoutFiles.setImgUrl(fileStorageService.uploadImage(musicRequest.getImg(), musicRequest.getAlbumName()));
+        } else {
+            musicWithoutFiles.setImgUrl(oldMusic.getImgUrl());
+        }
+        if (musicRequest.getSong() != null) {
+            fileStorageService.deleteMusicFile(oldMusic.getSongUrl());
+            musicWithoutFiles.setSongUrl(fileStorageService.uploadMusic(musicRequest.getSong(), musicRequest.getTestSongName()));
+        } else {
+            musicWithoutFiles.setSongUrl(oldMusic.getSongUrl());
+        }
+        musicRepository.updateMusic(musicWithoutFiles, id);
+    }
+
+    public List<Music> findAllMusicWithSearch(String search) {
+        var res = search.split(" ");
+        Set<Music> combinedSet = new LinkedHashSet<>();
+        for (String s : res) {
+            combinedSet.addAll(musicRepository.findAllMusicsWithSearch(s));
+        }
+        return new ArrayList<>(combinedSet);
+    }
+
+    @Transactional
+    public void changeCount(Long id, Integer count) {
+        musicRepository.changeCount(id, count);
+    }
+
+    public List<Music> findMusicSortedByPriceABS() {
+        return musicRepository.findAllSortedByPriceABSMusics();
+    }
+
+    public List<Music> findMusicSortedByPriceDESC() {
+        return musicRepository.findAllSortedByPriceDESCMusics();
+    }
+
+    public List<Music> findMusicSortedByReleaseDateABS() {
+        return musicRepository.findAllSortedByReleaseDateABSMusics();
+    }
+
+    public List<Music> findMusicSortedByReleaseDateDesc() {
+        return musicRepository.findAllSortedByReleaseDateDESCMusics();
     }
 
 }
